Add a command for writing mode and document functions better
Add a new function + command entry to enter writing mode in a tab. Only
works if there's one window and no more. Enables various settings that I
enjoy for writing and opens a split so that the text is centered nicely.

Also went through and documented my functions better, in case I need to
change/debug/understand them in the future.